chore: wait for extraction before asking if they want typescript or javascript
